+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [semantic versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added `save_csv_to_s3` function in `cdp/io/output.py`.
 
 ### Changed
 - Modified docstrings in `cdp/helpers/s3_utils.py`; remove type-hints